+++ src/test/java/de/pdv/apex/ExampleXML2PDFTest.java
@@ -88,7 +88,7 @@ class ExampleXML2PDFTest {
         try (PDDocument document = Loader.loadPDF(new RandomAccessReadBufferedFile(pdfFile))) {
             System.out.println("Pages: " + document.getNumberOfPages());
             System.out.println("Filesize (Bytes): " + Files.size(pdfFile.toPath()));
-            Assertions.assertEquals((int)1, document.getNumberOfPages());
+            Assertions.assertEquals(1, document.getNumberOfPages());
             System.out.println("Success!");
         }
     }
